fix: eagle3 quantized base model
